@@ -19,42 +19,59 @@
     "classnames": "^2.2.5",
     "cldr-data": "^30.0.2",
     "css-loader": "^0.25.0",
+    "dateformat": "^2.0.0",
     "formsy-react": "^0.18.1",
     "globalize": "^1.2.1",
+    "immutability-helper": "^2.1.2",
     "jquery": "^3.1.1",
     "jquery-ui": "^1.12.1",
-    "moment": "^2.16.0",
+    "moment": "^2.18.1",
     "newforms": "^0.13.2",
     "node": "0.0.0",
     "object-assign": "^4.1.0",
     "opentok-react": "^0.3.0",
     "pretty-checkbox": "^2.1.0",
+    "prop-types": "^15.5.8",
     "rc-time-picker": "^2.2.1",
-    "react": "^15.4.1",
+    "react": "^15.5.4",
+    "react-addons": "^0.9.1-deprecated",
+    "react-addons-test-utils": "^15.4.2",
     "react-big-calendar": "^0.11.1",
     "react-currency-input": "^1.1.1",
     "react-datepicker": "^0.37.0",
     "react-datetime": "^2.6.0",
     "react-dimensions": "^1.3.0",
-    "react-dom": "^15.4.1",
+    "react-dom": "^15.5.4",
+    "react-draggable": "^2.2.3",
     "react-dropzone": "^3.9.0",
     "react-dropzone-s3-uploader": "^0.10.0",
     "react-global": "^0.1.8",
+    "react-global-state": "^0.1.2",
     "react-iframe": "0.0.4",
     "react-maskedinput": "^3.3.4",
-    "react-measure": "^1.4.5",
+    "react-measure": "^1.4.6",
     "react-modal": "^1.5.2",
     "react-notification-system": "^0.2.11",
+    "react-onclickoutside": "^5.8.4",
+    "react-redux": "^5.0.4",
     "react-render-html": "^0.1.6",
+    "react-rnd": "^4.2.2",
     "react-rotating-text": "^1.1.2",
     "react-router": "^2.8.1",
+    "react-router-dom": "^4.1.1",
+    "react-scrollbar": "^0.5.1",
     "react-select": "^1.0.0-rc.2",
     "react-semantic-ui": "^0.2.0",
     "react-share": "^1.11.0",
+    "react-show-more": "^1.1.1",
     "react-text-truncate": "^0.8.3",
+    "react-textfit": "^0.1.8",
     "react-tinymce": "^0.5.1",
     "react-tinymce-input": "^1.2.1",
     "react-validated-input": "0.0.10",
+    "react-websocket": "^1.1.6",
+    "reconnecting-websocket": "^3.0.3",
+    "redux": "^3.6.0",
     "remarkable": "^1.6.2",
     "searchkit": "^0.10.1",
     "semantic-ui-daterangepicker": "^1.3.21",
