@@ -1,16 +1,14 @@
 from django.contrib import admin
-from kiterope.models import Profile, User, Goal, Student, Notification, Coach, Rate, Session, Review, Update, Post, Plan, Step, Question, Answer, Interest, Participant
-from kiterope.models import StepOccurrence, PlanOccurrence, UpdateOccurrence, Metric
-admin.site.register(Profile)
+from kiterope.models import Profile, User, Goal, Notification, Rate, Session, Review, Update, Post, Program, Step, Question, Answer, Interest, Participant
+from kiterope.models import StepOccurrence, PlanOccurrence, MessageThread, UpdateOccurrence, Metric, Message, Label, KChannel, KChannelUser, KRMessage, Contact
+
 admin.site.register(Goal)
-admin.site.register(Student)
-admin.site.register(Coach)
 admin.site.register(Rate)
 admin.site.register(Session)
 admin.site.register(Review)
 admin.site.register(Update)
 admin.site.register(Post)
-admin.site.register(Plan)
+admin.site.register(Program)
 admin.site.register(Step)
 admin.site.register(Question)
 admin.site.register(Answer)
@@ -21,6 +19,29 @@ admin.site.register(UpdateOccurrence)
 admin.site.register(Metric)
 admin.site.register(Participant)
 admin.site.register(Notification)
+admin.site.register(Message)
+admin.site.register(Label)
+admin.site.register(MessageThread)
+
+
+class KChannelUserInline(admin.TabularInline):
+    model = KChannelUser
+    extra = 1
+
+
+class KChannelAdmin(admin.ModelAdmin):
+    inlines = (KChannelUserInline,)
+
+
+admin.site.register(Profile)
+admin.site.register(KChannel, KChannelAdmin)
+admin.site.register(KChannelUser)
+
+admin.site.register(KRMessage)
+admin.site.register(Contact)
+
+
+
 
 
 
