@@ -24,9 +24,12 @@ from django.conf.urls import url
 from kiterope.views import schema_view
 from .views import ApiEndpoint
 import oauth2_provider.views as oauth2_views
+from django.conf.urls import include, url
 
 
+from rest_framework.authtoken.views import obtain_auth_token
 
+from django.views.decorators.csrf import csrf_exempt
 
 
 import apps
@@ -38,7 +41,8 @@ from rest_framework.urlpatterns import format_suffix_patterns
 
 router = routers.DefaultRouter()
 router.register(r'users', views.UserViewSet)
-router.register(r'plans', views.PlanViewSet)
+
+
 router.register(r'updates', views.UpdateViewSet)
 router.register(r'steps', views.StepViewSet)
 router.register(r'searchQuery', views.SearchQueryViewSet, 'SearchQuery')
@@ -46,18 +50,54 @@ router.register(r'goalEntry', views.GoalViewSet, base_name='Goal')
 router.register(r'sessions', views.SessionViewSet, base_name='Session')
 router.register(r'notifications', views.NotificationViewSet, base_name='Notification')
 
+router.register(r'messageThreads/(?P<thread_id>\w+)/messages', views.MessageThreadMessageViewSet, base_name='MessageThreadMessage')
+router.register(r'messageThreads/labels/(?P<label_id>\w+)', views.MessageThreadViewSet, base_name='MessageThread')
 
-router.register(r'^plans/(?P<plan_id>\w+)/steps', views.StepViewSet, base_name='Step')
-router.register(r'^period/(?P<periodRangeStart>[\w\-]+)/(?P<periodRangeEnd>[\w\-]+)', views.StepOccurrenceViewSet, base_name="StepOccurrence")
+router.register(r'messageThreads', views.MessageThreadViewSet, base_name='MessageThread')
+
+
+router.register(r'planOccurrences', views.PlanOccurrenceViewSet, base_name='PlanOccurrence')
+
+router.register(r'messages/(?P<sender_id>\w+)/(?P<receiver_id>\w+)', views.MessageViewSet, base_name='Message')
+router.register(r'messages', views.MessageViewSet, base_name='Message')
+router.register(r'labels/(?P<typeOfLabel>\w+)', views.LabelViewSet, base_name='Label')
+
+router.register(r'labels', views.LabelViewSet, base_name='Label')
+router.register(r'channels/(?P<channel_id>\w+)/messageThread', views.MessageThreadChannelViewSet, base_name='MessageThreadChannel')
+
+router.register(r'channels', views.KChannelViewSet, base_name='Room')
+
+router.register(r'channelUsers/(?P<receiver_id>\w+)', views.ReceiverKChannelViewSet, base_name='KChannel')
+router.register(r'^steps/(?P<step_id>\w+)/duplicate', views.StepDuplicatorViewSet, base_name='step-duplicate')
+
+router.register(r'^programs/(?P<program_id>\w+)/steps', views.StepViewSet, base_name='Step')
+router.register(r'^programs/(?P<program_id>\w+)/', views.ProgramViewSet, base_name='Program')
+router.register(r'programs', views.ProgramViewSet)
+
+
+router.register(r'^plans/(?P<plan_id>\w+)/steps', views.StepOccurrenceViewSet, base_name='Step')
+router.register(r'^plans', views.PlanOccurrenceViewSet, base_name='PlanOccurrence')
+
+
+router.register(r'^period/(?P<periodRangeStart>[\w\-]+)/(?P<periodRangeEnd>[\w\-]+)', views.PeriodViewSet, base_name="StepOccurrence")
 #router.register(r'period', views.GoalViewSet, base_name='Goal')
 router.register(r'^steps/(?P<step_id>\w+)/updates', views.UpdateViewSet, base_name='Update')
+
 router.register(r'goals', views.GoalViewSet, base_name='Goal')
-router.register(r'^goals/(?P<goal_id>\w+)/plans', views.PlanViewSet, base_name='Plan')
+router.register(r'contacts', views.ContactViewSet, base_name='Contact')
+
+router.register(r'^goals/(?P<goal_id>\w+)/planOccurrences', views.GoalPlanOccurrenceViewSet, base_name='PlanOccurrence')
+
+router.register(r'^goals/(?P<goal_id>\w+)/programs', views.ProgramViewSet, base_name='Program')
 router.register(r'^stepOccurrences/(?P<stepOccurrence_id>\w+)/updateOccurrences', views.UpdateOccurrenceViewSet, base_name='UpdateOccurrence')
-router.register(r'^myProfile', views.PersonalProfileViewSet)
+router.register(r'^stepOccurrences', views.StepOccurrenceViewSet, base_name='StepOccurrence')
+router.register(r'^updateOccurrences', views.UpdateOccurrenceViewSet, base_name='UpdateOccurrence')
+
+
+router.register(r'^clients', views.ClientViewSet, base_name='Profile')
 
 router.register(r'^profiles', views.ProfileViewSet, base_name='Profile')
-router.register(r'^plan/search', views.PlanSearchViewSet, base_name="plan-search")
+router.register(r'^program/search', views.ProgramSearchViewSet, base_name="program-search")
 
 
 
@@ -70,6 +110,7 @@ oauth2_endpoint_views = [
     url(r'^token/$', oauth2_views.TokenView.as_view(), name="token"),
     url(r'^revoke-token/$', oauth2_views.RevokeTokenView.as_view(), name="revoke-token"),
 ]
+
 
 if settings.DEBUG:
     # OAuth2 Application Management endpoints
@@ -90,55 +131,29 @@ if settings.DEBUG:
 
 urlpatterns = [
       url(r'^$', TemplateView.as_view(template_name='index.html')),
-      url(r'^interest', views.interest, name="interest"),
+      url(r'^', include('django.contrib.auth.urls')),
+
+                  # django-rest-auth (Login and registration via React)
+      url(r'^rest-auth/registration/', include('rest_auth.registration.urls')),
+      url(r'^rest-auth/', include('rest_auth.urls')),
+      url(r'^api/obtain-auth-token/$', obtain_auth_token),
       url(r'^swagger/', schema_view),
       url(r'^api/hello', ApiEndpoint.as_view()),  # an example resource endpoint
       url(r'^api/', include(router.urls)),
       url(r'^o/', include('oauth2_provider.urls', namespace='oauth2_provider')),
       url(r'^search/', include('haystack.urls')),
-
-
-                # url(r'^$', views.splash, name='splash'),
-
       url(r'^admin/', admin.site.urls),
       url(r'^accounts/', include('allauth.urls')),
-      url(r'^goals/add', views.goals_add, name='goals_add'),
       url(r'^secret', views.secret_page, name='secret'),
       url(r'^tinymce/', include('tinymce.urls')),
       url(r'^signS3Upload/$', views.sign_s3_upload, name='sign_s3_upload'),
-
-                  # REST FRAMEWORK URLS
-      # url(r'^users/$', views.UserList.as_view()),
-      # url(r'^users/(?P<pk>[0-9]+)/$', views.UserDetail.as_view()),
-      # url(r'^goals/$', views.GoalList.as_view()),
-      # url(r'^goals/(?P<pk>[0-9]+)$', views.GoalDetail.as_view()),
-      # url(r'^plans/$', views.PlanList.as_view()),
-      # url(r'^plans/(?P<pk>[0-9]+)$', views.PlanDetail.as_view()),
-      # url(r'^steps/$', views.StepList.as_view()),
-      # url(r'^steps/(?P<pk>[0-9]+)$', views.StepDetail.as_view()),
-      # url(r'^coaches/$', views.CoachList.as_view()),
-      # url(r'^coaches/(?P<pk>[0-9]+)$', views.CoachDetail.as_view()),
-      # url(r'^students/$', views.StudentList.as_view()),
-      # url(r'^students/(?P<pk>[0-9]+)$', views.StudentDetail.as_view()),
-      # url(r'^reviews/$', views.ReviewList.as_view()),
-      # url(r'^reviews/(?P<pk>[0-9]+)$', views.ReviewDetail.as_view()),
-
-      url(r'^rates/$', views.RateList.as_view()),
-      url(r'^rates/(?P<pk>[0-9]+)$', views.RateDetail.as_view()),
-
-      url(r'^updates/$', views.UpdateList.as_view()),
-      url(r'^updates/(?P<pk>[0-9]+)$', views.UpdateDetail.as_view()),
-
-      url(r'^questions/$', views.QuestionList.as_view()),
-      url(r'^questions/(?P<pk>[0-9]+)$', views.QuestionDetail.as_view()),
-
-      url(r'^answers/$', views.AnswerList.as_view()),
-      url(r'^answers/(?P<pk>[0-9]+)$', views.AnswerDetail.as_view()),
-
-      url(r'^goals/(?P<goal_id>\w+)/plans', views.goals_plans, name='goals_plans'),
       url(r'^api-auth/', include('rest_framework.urls')),
-      url(r'^plans/(?P<plan_id>\w+)', views.plan_edit, name='plan_edit'),
       url(r'^favicon.ico$', RedirectView.as_view(url='/static/favicon2.ico'), name="favicon"),
 
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
+if settings.DEBUG:
+    import debug_toolbar
+    urlpatterns += [
+        url(r'^__debug__/', include(debug_toolbar.urls)),
+    ]
